requirements.txt
asgiref==3.5.0
boto3==1.20.26
botocore==1.23.54
dj-database-url==0.5.0
django-auto-logout==0.5.0
django-ckeditor==6.0.0
django-ckeditor-5==0.1.6
django-crispy-forms==1.12.0
django-dotenv==1.4.2
django-environ==0.9.0
django-filter==21.1
django-formset-js-improved==0.5.0.2
django-heroku==0.3.1
django-jazzmin==2.4.8
django-jquery-js==3.1.1
django-js-asset==1.2.2
django-mailgun-provider==0.2.3
django-plaintext-password==0.1.0
django-rest-auth==0.9.5
django-rest-framework==0.1.0
django-social-share==2.2.1
django-static-fontawesome==5.14.0.0
django-storages==1.12.3
django-taggit==3.0.0
django-templated-mail==1.1.1
django-tinymce==3.4.0
django-widget-tweaks==1.4.8
djangorestframework==3.13.1
djangorestframework-simplejwt==5.2.0
djoser==2.0.5
gunicorn==20.0.4
lxml==4.6.2
pandas==1.4.0
Pillow==9.1.0
psycopg2==2.8.6
psycopg2-binary==2.9.1
python-decouple==3.5
requests==2.27.1
requests-oauthlib==1.3.1
rjsmin==1.1.0
s3transfer==0.5.2
sendgrid==6.9.7
uritemplate==4.1.1
urllib3==1.26.9
validate-email==1.3
whitenoise==5.2.0
django-import-export
shortuuid
django-shortuuidfield